chore: separate config package
