+++ packages/cli/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {},
-  "include": ["src", "rollup.config.ts"]
+  "include": ["src", "rollup.config.ts", "../config/src/check.ts"]
 }
