+++ packages/config/src/ScalarConfigJsonSchema.json
