+++ packages/config/src/check.test.ts
@@ -1,0 +1,17 @@
+import path from 'node:path'
+import { describe, expect, it } from 'vitest'
+
+import { check } from './check'
+
+describe('check', () => {
+  it('checks the valid json file', () => {
+    const result = check('/packages/config/src/valid.json')
+    console.log(JSON.stringify(result, null, 3))
+    expect(result.valid).toBe(true)
+  })
+  it('checks the invalid json file', () => {
+    const result = check('/packages/config/src/invalid.json')
+    console.log(JSON.stringify(result, null, 3))
+    expect(result.valid).toBe(true)
+  })
+})
