+++ packages/config/src/check.ts
