+++ packages/config/src/invalid.json
