+++ packages/config/src/valid.json
